requirements.txt
streamlit==1.45.1
tensorflow==2.18.0
Pillow
numpy
pandas
h5py